fix: 데이터 패칭 로직 분리
